backend/requirements.txt
Django
gunicorn
psycopg2-binary
django-cors-headers
djangorestframework
requests
djangorestframework-simplejwt
channels
channels_redis
python-socketio
aiohttp
uvicorn